+++ packages/protocol/contracts/L1/TaikoConfig.sol
@@ -32,7 +32,6 @@ library TaikoConfig {
             // of a geth transaction, so using 120KB for the proposed
             // transactions list calldata, 8K for the remaining tx fields.
             maxBytesPerTxList: 120000,
-            minTxGasLimit: 21000,
             proofCooldownPeriod: 30 minutes,
             systemProofCooldownPeriod: 15 minutes,
             // Only need 1 real zkp per 10 blocks.
@@ -42,7 +41,6 @@ library TaikoConfig {
             ethDepositGas: 21000,
             ethDepositMaxFee: 1 ether / 10,
             txListCacheExpiry: 0,
-            adjustmentQuotient: 16,
             relaySignalRoot: false
         });
     }