+++ packages/protocol/contracts/L1/libs/LibVerifying.sol
@@ -18,7 +18,7 @@ library LibVerifying {
     using SafeCastUpgradeable for uint256;
     using LibUtils for TaikoData.State;
 
-    event BlockVerified(uint256 indexed id, bytes32 blockHash);
+    event BlockVerified(uint256 indexed id, bytes32 blockHash, uint64 reward);
 
     event CrossChainSynced(uint256 indexed srcHeight, bytes32 blockHash, bytes32 signalRoot);
 
@@ -30,7 +30,8 @@ library LibVerifying {
         bytes32 genesisBlockHash,
         uint64 initBlockFee,
         uint64 initProofTimeTarget,
-        uint64 initProofTimeIssued
+        uint64 initProofTimeIssued,
+        uint16 adjustmentQuotient
     ) internal {
         if (
             config.chainId <= 1 || config.maxNumProposedBlocks == 1
@@ -38,14 +39,12 @@ library LibVerifying {
                 || config.blockMaxGasLimit == 0 || config.maxTransactionsPerBlock == 0
                 || config.maxBytesPerTxList == 0
             // EIP-4844 blob size up to 128K
-            || config.maxBytesPerTxList > 128 * 1024 || config.minTxGasLimit == 0
-                || config.maxEthDepositsPerBlock == 0
+            || config.maxBytesPerTxList > 128 * 1024 || config.maxEthDepositsPerBlock == 0
                 || config.maxEthDepositsPerBlock < config.minEthDepositsPerBlock
             // EIP-4844 blob deleted after 30 days
             || config.txListCacheExpiry > 30 * 24 hours || config.ethDepositGas == 0
                 || config.ethDepositMaxFee == 0 || config.ethDepositMaxFee >= type(uint96).max
-                || config.adjustmentQuotient == 0 || initProofTimeTarget == 0
-                || initProofTimeIssued == 0
+                || adjustmentQuotient == 0 || initProofTimeTarget == 0 || initProofTimeIssued == 0
         ) revert L1_INVALID_CONFIG();
 
         uint64 timeNow = uint64(block.timestamp);
@@ -55,6 +54,7 @@ library LibVerifying {
         state.blockFee = initBlockFee;
         state.proofTimeIssued = initProofTimeIssued;
         state.proofTimeTarget = initProofTimeTarget;
+        state.adjustmentQuotient = adjustmentQuotient;
         state.numBlocks = 1;
 
         TaikoData.Block storage blk = state.blocks[0];
@@ -66,7 +66,7 @@ library LibVerifying {
         fc.blockHash = genesisBlockHash;
         fc.provenAt = timeNow;
 
-        emit BlockVerified(0, genesisBlockHash);
+        emit BlockVerified(0, genesisBlockHash, 0);
     }
 
     function verifyBlocks(
@@ -114,7 +114,6 @@ library LibVerifying {
 
             _markBlockVerified({
                 state: state,
-                config: config,
                 blk: blk,
                 fcId: uint24(fcId),
                 fc: fc,
@@ -144,7 +143,6 @@ library LibVerifying {
 
     function _markBlockVerified(
         TaikoData.State storage state,
-        TaikoData.Config memory config,
         TaikoData.Block storage blk,
         TaikoData.ForkChoice storage fc,
         uint24 fcId,
@@ -158,7 +156,7 @@ library LibVerifying {
         uint64 reward = LibTokenomics.getProofReward(state, proofTime);
 
         (state.proofTimeIssued, state.blockFee) =
-            LibTokenomics.getNewBlockFeeAndProofTimeIssued(state, config, proofTime);
+            LibTokenomics.getNewBlockFeeAndProofTimeIssued(state, proofTime);
 
         unchecked {
             state.accBlockFees -= reward;
@@ -184,6 +182,6 @@ library LibVerifying {
         blk.nextForkChoiceId = 1;
         blk.verifiedForkChoiceId = fcId;
 
-        emit BlockVerified(blk.blockId, fc.blockHash);
+        emit BlockVerified(blk.blockId, fc.blockHash, reward);
     }
 }