+++ packages/protocol/script/DetermineNewProofTimeIssued.s.sol
@@ -5,11 +5,11 @@ import "forge-std/Script.sol";
 import "forge-std/console2.sol";
 import {LibLn} from "../test/LibLn.sol";
 
-uint16 constant DESIRED_PROOF_TIME_TARGET = 500;
-uint8 constant ADJUSTMENT_QUOTIENT = 16;
+uint16 constant DESIRED_PROOF_TIME_TARGET = 160;
+uint16 constant ADJUSTMENT_QUOTIENT = 32000;
 
 contract DetermineProofTimeIssued is Script {
-    function run() public {
+    function run() public view {
         uint16 proofTimeTarget = DESIRED_PROOF_TIME_TARGET; // Approx. value which close to what is in the simulation
         uint64 feeBase = 1e8; // 1 TKO
         uint64 initProofTimeIssued =