+++ packages/status-page/src/utils/getNumProvers.ts
@@ -14,6 +14,7 @@ export const getNumProvers = async (
   const uniqueProversResp = await axios.get<UniqueProverResponse>(
     `${eventIndexerApiUrl}/uniqueProvers`
   );
+
   if (uniqueProversResp.data) {
     uniqueProversResp.data.provers.sort((a, b) => b.count - a.count);
     // Filter out the oracle prover address since it doesn't submit the actual zk proof
@@ -25,6 +26,16 @@ export const getNumProvers = async (
       uniqueProversResp.data.provers.splice(index, 1);
       uniqueProversResp.data.uniqueProvers--;
     }
+    // Filter out the system prover address since it doesn't submit the actual zk proof
+    const systemIndex = uniqueProversResp.data.provers.findIndex(
+      (uniqueProver) =>
+        uniqueProver.address === "0x0000000000000000000000000000000000000001"
+    );
+    if (systemIndex > -1) {
+      uniqueProversResp.data.provers.splice(systemIndex, 1);
+      uniqueProversResp.data.uniqueProvers--;
+    }
   }
+
   return uniqueProversResp.data || { uniqueProvers: 0, provers: [] };
 };