+++ packages/status-page/src/utils/initConfig.ts
@@ -2,12 +2,12 @@ import { ethers } from "ethers";
 import { Layer } from "../domain/layer";
 
 export function initConfig(layer: Layer) {
-  const l1Provider = new ethers.providers.JsonRpcProvider(
+  const l1Provider = new ethers.providers.StaticJsonRpcProvider(
     layer === Layer.Two
       ? import.meta.env.VITE_L1_RPC_URL
       : import.meta.env.VITE_L2_RPC_URL
   );
-  const l2Provider = new ethers.providers.JsonRpcProvider(
+  const l2Provider = new ethers.providers.StaticJsonRpcProvider(
     layer === Layer.Two
       ? import.meta.env.VITE_L2_RPC_URL
       : import.meta.env.VITE_L3_RPC_URL
@@ -21,6 +21,7 @@ export function initConfig(layer: Layer) {
     layer === Layer.Two
       ? import.meta.env.VITE_L2_TAIKO_L2_ADDRESS
       : import.meta.env.VITE_L3_TAIKO_L2_ADDRESS;
+  const taikoTokenAddress = import.meta.env.VITE_TAIKO_TOKEN_ADDRESS;
   const l1ExplorerUrl = import.meta.env.VITE_L1_EXPLORER_URL;
   const l2ExplorerUrl =
     layer === Layer.Two
@@ -39,6 +40,7 @@ export function initConfig(layer: Layer) {
     l2Provider,
     l1TaikoAddress,
     l2TaikoAddress,
+    taikoTokenAddress,
     l1ExplorerUrl,
     l2ExplorerUrl,
     feeTokenSymbol,